+++ extensions/threads/src/lib/download-media.ts
@@ -2,7 +2,7 @@ import axios from "axios";
 import { createWriteStream, existsSync } from "fs";
 import { showToast, Toast } from "@raycast/api";
 
-export async function getThreadsVideoURL(threadsUrl: string) {
+export async function getThreadsMediaURL(threadsUrl: string) {
   try {
     const response = await axios.get(
       `https://api.threadsphotodownloader.com/v2/media?url=${threadsUrl}`,
@@ -14,9 +14,13 @@ export async function getThreadsVideoURL(threadsUrl: string) {
       },
     );
 
+    const imageUrls = response.data["image_urls"];
     const videoUrls = response.data["video_urls"];
 
-    return videoUrls;
+    return {
+      images: imageUrls,
+      videos: videoUrls,
+    };
   } catch (error) {
     if (axios.isAxiosError(error)) {
       console.error("Error:", error.message);
@@ -28,28 +32,29 @@ export async function getThreadsVideoURL(threadsUrl: string) {
 }
 
 export async function handleDownload(
-  videoUrl: string,
-  videoId: string,
+  mediaUrl: string,
+  mediaId: string,
   downloadFolder: string,
+  fileExtension: string,
 ) {
-  let filePath = `${downloadFolder}/${videoId.substring(0, 100)}.mp4`;
+  let filePath = `${downloadFolder}/${mediaId.substring(0, 100)}.${fileExtension}`;
   let counter = 1;
 
   while (existsSync(filePath)) {
-    filePath = `${downloadFolder}/${videoId.substring(0, 100)}(${counter}).mp4`;
+    filePath = `${downloadFolder}/${mediaId.substring(0, 100)}(${counter}).${fileExtension}`;
     counter++;
   }
 
   const writer = createWriteStream(filePath);
 
   const progressToast = await showToast({
-    title: "Downloading Video",
+    title: "Downloading Media",
     message: "0%",
     style: Toast.Style.Animated,
   });
 
   try {
-    const response = await axios.get(videoUrl, {
+    const response = await axios.get(mediaUrl, {
       responseType: "stream",
       onDownloadProgress: (event) => {
         if (event.total) {
@@ -68,12 +73,12 @@ export async function handleDownload(
 
     await showToast({
       title: "Download Complete",
-      message: `Video saved to ${filePath}`,
+      message: `Media saved to ${filePath}`,
       style: Toast.Style.Success,
     });
   } catch (error) {
     await showToast({
-      title: "Error While Downloading Video",
+      title: "Error While Downloading Media",
       message:
         error instanceof Error ? error.message : "Unknown error occurred",
       style: Toast.Style.Failure,