+++ extensions/threads/CHANGELOG.md
@@ -2,6 +2,6 @@
 
 ## [Feature] - {PR_MERGE_DATE}
 
-- Added the `Download Threads Video` command to download videos from Threads.
+- Added the `Download Threads Media` command to download medias from Threads.
 
 ## [Initial Version] - 2024-08-21